Fixed the following bugs: - UnitCount only accepted ints (now accepts floats) - Multiple radio buttons could be selected (now selecting a new one unchecks the old one) - Unit type is actually sent to server now instead of "none" placeholder - Got rid of floating mystery button on Login screen

+++ app/src/main/java/net/azurewebsites/athleet/ApiLib/ApiProvider.kt
@@ -42,7 +42,7 @@ interface Api {
         @Query("DefaultReps") defaultReps:Int,
         @Query("ExerciseSets") exerciseSets: Int,
         @Query("MeasureUnits") measureUnits:String,
-        @Query("unitCount") unitCount: Int,
+        @Query("unitCount") unitCount: Float,
         @Query("WorkoutID") workoutID: Int
     ):Call<ResponseBody>
